fix: run go mod tidy after scaffold
