+++ starport/services/scaffolder/init.go
@@ -9,10 +9,6 @@ import (
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/gobuffalo/genny"
-	conf "github.com/tendermint/starport/starport/chainconf"
-	"github.com/tendermint/starport/starport/pkg/cosmosanalysis/module"
-	"github.com/tendermint/starport/starport/pkg/cosmosgen"
-	"github.com/tendermint/starport/starport/pkg/giturl"
 	"github.com/tendermint/starport/starport/pkg/gomodulepath"
 	"github.com/tendermint/starport/starport/pkg/localfs"
 	"github.com/tendermint/starport/starport/pkg/placeholder"
@@ -48,13 +44,7 @@ func (s *Scaffolder) Init(tracer *placeholder.Tracer, name string) (path string,
 		return "", err
 	}
 
-	// generate protobuf types
-	if err := s.protoc(absRoot, pathInfo.RawPath); err != nil {
-		return "", err
-	}
-
-	// format the source
-	if err := fmtProject(absRoot); err != nil {
+	if err := s.finish(absRoot, pathInfo.RawPath); err != nil {
 		return "", err
 	}
 
@@ -106,47 +96,6 @@ func (s *Scaffolder) generate(tracer *placeholder.Tracer, pathInfo gomodulepath.
 	// generate the vue app.
 	vuepath := filepath.Join(absRoot, "vue")
 	return localfs.Save(vue.Boilerplate(), vuepath)
-}
-
-func (s *Scaffolder) protoc(projectPath, gomodPath string) error {
-	if err := cosmosgen.InstallDependencies(context.Background(), projectPath); err != nil {
-		return err
-	}
-
-	confpath, err := conf.LocateDefault(projectPath)
-	if err != nil {
-		return err
-	}
-	conf, err := conf.ParseFile(confpath)
-	if err != nil {
-		return err
-	}
-
-	options := []cosmosgen.Option{
-		cosmosgen.WithGoGeneration(gomodPath),
-		cosmosgen.IncludeDirs(conf.Build.Proto.ThirdPartyPaths),
-	}
-
-	// generate Vuex code as well if it is enabled.
-	if conf.Client.Vuex.Path != "" {
-		storeRootPath := filepath.Join(projectPath, conf.Client.Vuex.Path, "generated")
-
-		options = append(options,
-			cosmosgen.WithVuexGeneration(
-				false,
-				func(m module.Module) string {
-					parsedGitURL, _ := giturl.Parse(m.Pkg.GoImportName)
-					return filepath.Join(storeRootPath, parsedGitURL.UserAndRepo(), m.Pkg.Name, "module")
-				},
-				storeRootPath,
-			),
-		)
-	}
-	if conf.Client.OpenAPI.Path != "" {
-		options = append(options, cosmosgen.WithOpenAPIGeneration(conf.Client.OpenAPI.Path))
-	}
-
-	return cosmosgen.Generate(context.Background(), projectPath, conf.Build.Proto.Path, options...)
 }
 
 func initGit(path string) error {
