+++ starport/services/scaffolder/query.go
@@ -70,8 +70,5 @@ func (s *Scaffolder) AddQuery(
 	if err != nil {
 		return err
 	}
-	if err := s.protoc(pwd, path.RawPath); err != nil {
-		return err
-	}
-	return fmtProject(pwd)
+	return s.finish(pwd, path.RawPath)
 }
